test(UI): Successful execusion of React Boilerplate tests in UI folder
