+++ ui/.eslintrc
@@ -1,10 +1,8 @@
 {
+	"extends": "airbnb",
 	"rules": {
-		"import/no-named-as-default": 0,
-		"import/no-named-as-default-member": 0,
-		"react/require-default-props": 0,
-		"react/forbid-prop-types": 0,
-		"jsx-a11y/label-has-associated-control": 0,
-		"jsx-a11y/label-has-for": 0
-	}
+		"max-len": [2, 150, 4],
+		"jsx-a11y/tabindex-no-positive": "off"
+	},
+	"parser": "babel-eslint"
 }
