+++ ui/app/components/Footer/tests/index.test.js
@@ -26,7 +26,7 @@ describe('<Footer />', () => {
           <FormattedMessage
             {...messages.authorMessage}
             values={{
-              author: <A href="https://twitter.com/mxstbr">Max Stoiber</A>,
+              author: <A href="https://twitter.com/danactive">Dan BROOKS</A>,
             }}
           />
         </section>,
