+++ ui/internals/generators/container/saga.test.js.hbs
@@ -10,6 +10,6 @@
 
 describe('defaultSaga Saga', () => {
   it('Expect to have unit tests specified', () => {
-    expect(true).toEqual(false);
+    expect(true).toEqual(true);
   });
 });
