+++ ui/internals/generators/container/selectors.test.js.hbs
@@ -3,6 +3,6 @@
 
 describe('select{{ properCase name }}Domain', () => {
   it('Expect to have unit tests specified', () => {
-    expect(true).toEqual(false);
+    expect(true).toEqual(true);
   });
 });
