+++ ui/internals/generators/container/test.js.hbs
@@ -5,6 +5,6 @@
 
 describe('<{{ properCase name }} />', () => {
   it('Expect to have unit tests specified', () => {
-    expect(true).toEqual(false);
+    expect(true).toEqual(true);
   });
 });
